power: smb5: Add PD HV Support
Add the ability to properly handle PD chargers
when configured higher then 9 V.  This change
will enable the Voltage Max to be set and then
PDO selection to be done.

The ICL and the Buck Frequency also needed to
be set properly as well as some rework to
prevent spamming calls.

refer to
https://gerrit.mot.com/969310
https://gerrit.mot.com/969311
https://gerrit.mot.com/969312
https://gerrit.mot.com/1085408
https://gerrit.mot.com/1085409

Change-Id: Ied7f5c66057c7431bd65049291a9521bdeadcc0c
Signed-off-by: Haijian Ma <mahj8@motorola.com>
Reviewed-on: https://gerrit.mot.com/1262530
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key