power: smb5-lib: Fix Non-Default CC ICL
Current implementation will ignore the CC line Rp
resistance if the BC 1.2 detection is PC Port
and the USB driver sends 500/900 mA as the Current
limit.  To fix this only listen to the USB driver
if the CC Rp is Default.

Also a secondary defect was found because 900
mA was not properly being set from a USB 3.1 port.
To fix that ensure the the cl_usb is checked to see
that it is at or below 500 mA prior to locking 500 mA.

refer to https://gerrit.mot.com/1026470

Change-Id: Icb1323379d5701f02f80073a7346dabcd0c08f7a
Signed-off-by: Haijian Ma <mahj8@motorola.com>
Reviewed-on: https://gerrit.mot.com/1281457
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key